Add persistence measurement as command line argument.

+++ common/persistence.c
@@ -120,7 +120,7 @@ int _persistence_init (persistence_agent_t *agent, const char *filename)
     return 0;
 }
 
-persistence_agent_t *persistence_init (const char *filename, uint8_t flags)
+persistence_agent_t *persistence_init (const char *filename, uint32_t flags)
 {
     persistence_agent_t *agent = malloc (sizeof (persistence_agent_t));
     agent->flags = flags;
